rgw: Allow to serve Swift off the URL root
In most public clouds OpenStack Swift URL is based on the tree
root, like:
  https://swift.domain.com/v1/AUTH_tenant

Our convention is to have a path, like:
  https://domain.com/swift/v1/AUTH_tenant

The path is configurable, but even if someone manages to set
an empty prefix, we print headers with %s/%s/v1 where the extra
slash is baked in. In addition, the empty prefix means "use the
/swift default".

To provide compatibility, this patch uses a trick. The empty
prefix remains to mean "/swift", but then we use "/" (one slash)
to create the needed empty prefix.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>