doc: amend Fixes instructions in SubmittingPatches
Since the "Fixes: #..." line is interpreted by GitHub as referring to a
pull request, yet the intention is for it to refer to the tracker issue,
change our instructions to use the full tracker issue URL.

Now that github pull request numbers are in the 8000s, there are cases when
a PR is wrongly closed by a "Fixes: #..." line in a commit message. See

https://github.com/ceph/ceph/pull/8286#event-603120191

for one such case.

Signed-off-by: Nathan Cutler <ncutler@suse.com>