messages/MOSDOp: clear reqid inc for v6 encoding
For the new v7 encoding, we put the client incarnation in the
reqid so that we can have it at the beginning of the encoding.
However, we do *not* want to do this for the v6 encoding, as that
version of get_reqid() will return reqid if it is not ==
osd_reqid_t().

Fixes: #15230
Signed-off-by: Sage Weil <sage@redhat.com>