osd/OSDMap: fix clean_temps and test
For both pg_temp and primary_temp, we want to calculate what the
mapping *will* be after the inc is applied, and if the mapping is
redundant (it matches what we would have gotten anyway), remove
it.  Removing the mapping might mean putting a 'remove' entry in
the inc, or it might mean removing the proposed addition in the
inc.

Update the unit test to test both cases.

Signed-off-by: Sage Weil <sage@redhat.com>