kv: explicitly set DISABLE_JEMALLOC for rocksdb in Makefile.am.
we should disable rocksdb from using jemalloc even it is installed
in the building host. as it causes unresolved symbols if ceph is not
configured to use jemalloc and hence is not linked against it.

we can pass "DISABLE_JEMALLOC=0" to rocksdb's Makefile in hope to enable
jemalloc. because rocksdb checks for jemalloc using
```
ifndef DISABLE_JEMALLOC
```

Signed-off-by: Kefu Chai <kchai@redhat.com>