librbd: introduce RBD_IMAGE_OPTION_FEATURES_[SET|CLEAR] for features overwriting.
Currently, if we want to use the default options for rbd, we need to omit the
RBD_IMAGE_OPTION_FEATURES, but if we want --image-shared. we need to overwrite
something bese on the default value of image options.

This patch introduce two flags in image_options, RBD_IMAGE_OPTION_FEATURES_SET
means, we want to set something after you get the features from default, parent
or user. And RBD_IMAGE_OPTION_FEATURES_CLEAR will do a bit clear.

Signed-off-by: Dongsheng Yang <dongsheng.yang@easystack.cn>