rpm: proper packaging of udev directories
The issue here is that ceph.spec.in does not package the directories
/usr/lib/udev and /usr/lib/udev/rules.d. The problem was not showing because
hdparm, which is brought in as a build dependency, packages these directories.
However, in SUSE a recent update to hdparm changes that and the problem
manifests.

This PR addresses the issue by adding udev as a build dependency, which should
cover /usr/lib/udev, and by adding explicit "%dir %{_udevrulesdir}" to the
packages that put files in this directory.

Fixes: http://tracker.ceph.com/issues/16949
Signed-off-by: Nathan Cutler <ncutler@suse.com>