suites/kcephfs: put clients on dedicated nodes
Some tests want to use power cycling to reset stuck
mounts, and that needs to not kill OSDs as collateral
damage.

Need to revisit this to avoid unnecessarily using a whole
node for the client for those tests that don't require it
(i.e. those that don't use CephFSTestCase.REQUIRE_KCLIENT_REMOTE)

Signed-off-by: John Spray <john.spray@redhat.com>