MOSDMap: re-encode if CEPH_FEATURE_OSD_HITSET_GMT is not set
re-encode MOSDMap with the connection's features instead of using the
features of quorum. prior to this change, we always encode incremental
osdmap using the quorum features, but on the OSD side, it could be on
an older version, so may not be able to re-encode the full map using
the quorum features, hence the resulting CRC won't match with the
one sent along with the incremental map from monitor. with this change,
we re-encode OSDMaps if the connection features (the intersection of
quorum features and the peer features) does not have the
CEPH_FEATURE_OSD_HITSET_GMT feature bit. because the re-encoded osdmaps
(both inc and full) are not cached in the leveldb, every time an OSDMap
is sent, we reencode it if CEPH_FEATURE_OSD_HITSET_GMT is not set. this
would degrade the performance of monitor, but the less old client
without HITSET_GMT support in the cluster, the less performance
degradation is expected.

Fixes: http://tracker.ceph.com/issues/17386
Signed-off-by: Kefu Chai <kchai@redhat.com>

Resolves: rhbz#1378549