os/bluestore: remove deferred_csum machinery
When we added this way back in 36e0a4b5cc21c8d7a969d7e81c9a521770d5691e,
we did not have our own buffer cache, and were relying
on the cache at the BlockDevice layer.  In that case,
we would have the problem of a partial wal overwrite
followed by another partial write that needed to read
the rest of the chunk.

However, now we have our own cache, and any data we write
in the _do_write_small() wal path will go into the cache,
which means we will never read the old data off of
disk and need the old csum values.

Remove this now-unnecessary kludge!

Signed-off-by: Sage Weil <sage@redhat.com>