mds: fix false 'bad backtrace' warnings during online scrub
Only inode's direct parent in backtrace is supposed to be uptodate.
The rest back pointers in backtrace are hints, they can be stale.
The scrub code prints warning when inode's on-disk backtrace is
not equivalent to inode's current path. This check is too strange,
it causes lots of false warnings for workload that does directory
rename.

Fixes: http://tracker.ceph.com/issues/17562
Signed-off-by: Yan, Zheng <zyan@redhat.com>