doc: fix mgr literal block rST syntax
Prior to this change, there were double-colon ("::") characters in the
HTML docs to indicate literal blocks.

The proceeding paragraphs were not pre-formatted because there was no
newline between the double-colon and the to-be-pre-formatted paragraph.

Move the "::" characters inline to simplify the syntax and make Sphinx
properly interpret these texts as literal blocks.

Signed-off-by: Ken Dreyer <kdreyer@redhat.com>