rgw: pass authentication domain to civetweb
This is needed to support absolute uris, as civetweb further cross
checks whether the request is made to the same domain or not, otherwise
civetweb ends up sending NULL response back to the client. Also set the
REQUEST_URI to the full request uri, since the old uri parameter is
deprecated.

Fixes: http://tracker.ceph.com/issues/17657
Signed-off-by: Abhishek Lekshmanan <abhishek@suse.com>