OSD, PG: remove optional arg from queue_for_recovery
The optional arg ("front") was meant to control if PG was supposed
to be put in front or back (default) of awaiting_throttle. For some
reason, this was't used at all, so this commit removes it and
replaces with logic that checks whether the PG has forcecd
backfill or recovery set, and lets it in the front only in that
case.

Signed-off-by: Piotr Dałek <piotr.dalek@corp.ovh.com>