Revert "Revert "ceph-disk: change get_dmcrypt_key test to support different cluster name""
This reverts commit 9c776efb19e4416e7a89454fd6692a49a2e4395c which
reverted 6fb8e87afdb5bc6a6913acb3a6f8c2d163e2b7ac. Although in the same
pull request, 6fb8e87afdb5bc6a6913acb3a6f8c2d163e2b7ac was a fix for
45359c9abfc559827db96f6d7c45cb124d47e9fc. The two should have been
squashed but they were not.

When reverting the pull request I reverted all the commits in the pull
request but I did it in the wrong order and the fix became the
problem. Luckily the outcome is a noop instead of a bug, reason why
it was undetected by the tests.

Signed-off-by: Loic Dachary <ldachary@redhat.com>