mds: set ceph-mds name uncond for external tools
External tools like logrotate which use killall rely on the "ceph-mds" name to
find it in /proc/*/comm. This is normally the case but when ceph-mds respawns
using /proc/self/exe (on Linux), its name will change to "exe". This makes
logrotate fail to signal ceph-mds the log needs to be reopened and will lead to
the log growing until the disk is full.

This patch unconditionally sets the name so on respawn the name is correct (and
in any other possible scenario, such as a misnamed executable).  Note, there is
still a very small race between execve and ceph_pthread_setname where the
process name is wrong.

Problem was introduced by 590c1d4719efdde417eca80dc318b3dc04e8d8a0.

Fixes: http://tracker.ceph.com/issues/19291

Signed-off-by: Patrick Donnelly <pdonnell@redhat.com>