common/ceph_context: fewer warnings about experimental features
As Pete Zaitcev points out, check_experimental_feature_enabled
warns when an actual experimental feature is used; we don't need
to also warn that we might use it here.

This means we don't notice when a feature is enabled but it isn't
used, but that noise arguably has no value.

Signed-off-by: Sage Weil <sage@redhat.com>